Window Maximize and New Feature Files
I made the following additions: 1) Testing is now done with maximized windows. This is to prevent Selenium failing to find elements because they are hidden due to mobile adjustments; 2) Added or appended to the AddCats, FilterCats, and SearchCats feature files; 3) CatDAO adjustment to queries to ensure cats haven't been adopted/purchased.